feat: release 6.8.0 with reliable GPU defaults
Default model execution to Precision.fp32 and update flutter_litert to 3.8.0, while retaining opt-in fp16 support. Optimize multiclass mask handling and refresh the example Android toolchain for Flutter beta CI.

+++ lib/src/isolate/face_detector_core.dart
@@ -126,7 +126,7 @@ class _FaceDetectorCore {
     int meshPoolSize = 3,
     bool useCompiledModel = false,
     Set<Accelerator> accelerators = const {Accelerator.gpu, Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
     double minScore = 0.0,
     double minFaceSize = 0.0,
     double minFacePresenceConfidence = 0.0,
